Initial commit: add first files
This first commit has quite a few things:

* A begginer's Makefile. It requires some polish that may already be
present in the original, but got deleted when I started poking around.
* gosh-main.go has the flag-parsing and calling interface facilities.
* gosh.go contains the actual shell implementation.  I structured it in
three main functions: a main, Gosh(), and two goroutines to read from
stdin and execute the commands, respectively.

Right now all it does is echo the command-line back to the user.